Refactoring to models package
