+++ backend/api/models/_tests/test_metrics.py
@@ -6,7 +6,7 @@ import pytest
 from dateutil.relativedelta import relativedelta
 
 from api._tests.test_fixtures import generate_expected_timeline
-from api.metrics import InstallActivity
+from api.models.metrics import InstallActivity
 
 PLUGIN_NAME = 'foo'
 
@@ -43,7 +43,7 @@ class TestInstallActivity:
         expected = 173
         mock_install_activity = Mock(return_value=Mock(install_count=expected))
 
-        from api.metrics import InstallActivity
+        from api.models.metrics import InstallActivity
         monkeypatch.setattr(InstallActivity, 'get', mock_install_activity)
         actual = InstallActivity.get_total_installs(PLUGIN_NAME)
 
@@ -54,7 +54,7 @@ class TestInstallActivity:
         def raise_exception(_, __):
             raise InstallActivity.DoesNotExist()
 
-        from api.metrics import InstallActivity
+        from api.models.metrics import InstallActivity
         monkeypatch.setattr(InstallActivity, 'get', raise_exception)
         actual = InstallActivity.get_total_installs(PLUGIN_NAME)
 
@@ -69,7 +69,7 @@ class TestInstallActivity:
         mock_install_activity = Mock(return_value=results)
         day_delta = 15
 
-        from api.metrics import InstallActivity
+        from api.models.metrics import InstallActivity
         monkeypatch.setattr(InstallActivity, 'query', mock_install_activity)
         actual = InstallActivity.get_recent_installs(PLUGIN_NAME, day_delta)
 
@@ -86,7 +86,7 @@ class TestInstallActivity:
     def test_get_timeline(self, monkeypatch, results, month_delta, expected):
         mock_install_activity = Mock(return_value=results)
 
-        from api.metrics import InstallActivity
+        from api.models.metrics import InstallActivity
         monkeypatch.setattr(InstallActivity, 'query', mock_install_activity)
         actual = InstallActivity.get_timeline(PLUGIN_NAME, month_delta)
